// src/app/services/models/receipt.ts
export class Receipt {
    id: string;
    no: string;
    start: string;
    end: string;
    total: string;
    item: Array<any>;
    posID: string;
    posName: string;
    posNumber: string;
    storeID: string;
    storeName: string;
    storeCode: string;
    receiptType: string;
    fullname: string;
    type: string;
}